This adds Wingnut's auto-credit patch (except it doesn't use a separate auto_credit flag; that's what sell_response is for).
follow-up comment from Pat:
     This introduces behavior that I consider to be buggy.
You need something separate from sell_response, otherwise if
you answer (a)ll when the shopkeeper still has money you end
up selling everything for credit without being asked when he
runs out of cash in the middle of the transation.  Avoiding
that is the reason for the old behavior in the first place.