do_look at self
     Suggested by <Someone>:  when examining an `@' monster, only include
"or you" in the description if you might reasonably expect to be seen
as something else.  So if you're playing as a human or an elf, or if
you're polymorphed regardless of race, the ordinary description of
"human or elf" doesn't need to add "or you".  And the setting of the
`showrace' option only matters if the `@' comes from prompting the
user instead of picking a symbol from the map.