vms build fix
     Noticed when trying out the "heck patch", where the name of the file
for the sanctum level was changed and I neglected to update my Makefile to
reflect that:  nothing was noticed if any files were missing when loading
them into a dlb container.  Populating the playground directory for the
non-dlb configuration suffered a similar problem.  Now the VMS playground
setup will issue a warning message if this happens (but not abort the
installation; perhaps it ought to do that as well...).

     Other ports have the same problem.  For example, Unix Makefile.top
relies on file globbing to build the dlb container; if a wildcard pattern
matches no files, dlb won't notice because the shell will have eaten the
offending pattern.  Likewise for its non-dlb configuration, where cp won't
notice.  However, this isn't a very interesting bug because it won't hit
when the distributed Makefiles correspond with the needed data files.